Dialogue Zones and Pause Update
DialogTrigger now has two settings: IMMEDIATE and ON_PLAYER_ENTER.

IMMEDIATE DialogTriggers will instantly attempt to send their dialogue to the UI, such as on the main menu.

ON_PLAYER_ENTER DialogTriggers will only send their dialogue to the UI when a player enters the corresponding BoxCollider.

DialogManager now has references to all of the subcomponents of "Dialog" in the canvas.  This allows it to enable or disable them from view, using the new ToggleAllDialogUI(bool) method. It also has NextButton which will force DisplayNextSentence() when clicked.

UI is hidden by default now, and the canvas prefab has been updated for all levels so that it only displays when dialogue is sent to the UI, and automatically hides when there is no dialogue left to read.

PAUSE_GAME has been updated to be made of two component booleans: PauseButton and DialogPause. The game will pause itself if EITHER the player pauses, or the dialogue system pauses. This distinction also prevents the player from unpausing the game by pressing the pause button when the dialogue system has forcefully paused the game!